leaves/serializers.py
import logging
from datetime import date
from .models import Leave, Employee, LeaveType, Institution
from .utils import calculate_working_days, send_welcome_email
from rest_framework import serializers
from django.contrib.auth.tokens import default_token_generator
from django.utils.http import urlsafe_base64_decode
from django.utils.encoding import force_str
from django.utils import timezone

logger = logging.getLogger(__name__)


# Serializer for Institution model
class InstitutionSerializer(serializers.ModelSerializer):
    employee_count = serializers.SerializerMethodField()
    class Meta:
        model = Institution
        fields = ["id", "name", "location", "created_at", "is_active", "employee_count"]
    
    def get_employee_count(self, obj):
        return obj.employees.filter(is_active=True).count()

# Serializer for Employee model
class EmployeeSerializer(serializers.ModelSerializer):
    leave_count = serializers.SerializerMethodField()
    institution_name = serializers.SerializerMethodField()

    class Meta:
        model = Employee
        fields = [
            "id",
            "email",
            "first_name",
            "last_name",
            "department",
            "position",
            "role",
            "institution",
            "institution_name",
            "leave_count",
            "must_reset_password",
            "is_active",
        ]

    def get_institution_name(self, obj):
        return obj.institution.name if obj.institution else None
    
    def get_leave_count(self, obj):
        return obj.leaves.count()
        
    def update(self, instance, validated_data):
        password = validated_data.pop("password", None)

        for attr, value in validated_data.items():
            setattr(instance, attr, value)
        if password:
            instance.set_password(password)
        instance.save()
        return instance
    

class EmployeeCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = [
            "id",
            "email",
            "first_name",
            "last_name",
            "department",
            "position",
            "role",
            "institution",
            "phone_number",
        ]
        extra_kwargs = {
            "password": {"write_only": True},
        }

    def create(self, validated_data):
        password = validated_data.pop("password", None)
        employee = Employee(**validated_data) 
        employee.set_unusable_password()
        employee.must_reset_password = True
        employee.save()

        send_welcome_email(employee)
        logger.info(f"Created new employee: {employee.email} and sent welcome email.")

        return employee


class EmployeeUpdateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = [
            "email",
            "first_name",
            "last_name",
            "department",
            "position",
            "role",
            "institution",
            "phone_number",
            "must_reset_password",
            "is_active",
        ]

class SetPasswordSerializer(serializers.Serializer):
    """Serializer for allowing users to set a new password after receiving a reset link."""
    uid = serializers.CharField()
    token = serializers.CharField()
    new_password = serializers.CharField(min_length=8,write_only=True, required=True)
    confirm_password = serializers.CharField(min_length=8, write_only=True, required=True)
    

    def validate_password(self, value):
        if value['new_password'] != value['confirm_password']:
            raise serializers.ValidationError("Passwords do not match.")
        if len(value['new_password']) < 8:
            raise serializers.ValidationError("Password must be at least 8 characters long.")

        try:
            uid = force_str(urlsafe_base64_decode(value['uid']))
            employee = Employee.objects.get(pk=uid)
        except(Employee.DoesNotExist, ValueError, TypeError, OverflowError):
            raise serializers.ValidationError({"uid": "Invalid user."})
        
        if not default_token_generator.check_token(employee, value['token']):
            raise serializers.ValidationError({"token": "Reset link is invalid or has expired."})
        
        value['employee'] = employee
        return value
    
    def save(self):
        employee = self.validated_data["employee"]
        employee.set_password(self.validated_data["new_password"])
        employee.must_reset_password = False
        employee.save()
        logger.info(f"User {employee.email} has reset their password.")

        return employee


# Login Serializer
class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)

class LeaveTypeSerializer(serializers.ModelSerializer):
    leave_count = serializers.SerializerMethodField()

    class Meta:
        model = LeaveType
        fields = ["id", "name", "description", "leave_count"]
    
    def get_leave_count(self, obj):
        return obj.leaves.count()
    
    def  validate_max_days(self, value):
        if value is not None and value <= 0:
            raise serializers.ValidationError("Max days must be a positive integer.")
        return value

class LeaveSerializer(serializers.ModelSerializer):
    employee_name = serializers.CharField(source="employee.get_full_name", read_only=True)
    leave_type_name = serializers.CharField(source="leave_type.name", read_only=True)
    institution_name = serializers.CharField(source="employee.institution.name", read_only=True)
    leave_duration = serializers.SerializerMethodField()

    class Meta:
        model = Leave
        fields = [
            "id",
            "employee",
            "employee_name",
            "leave_type",
            "leave_type_name",
            "institution_name",
            "start_date",
            "end_date",
            "reason",
            "status",
            "admin_remarks",
        ]

    def get_leave_duration(self, obj):
        return calculate_working_days(obj.start_date, obj.end_date)
    
    def get_employee_name(self, obj):
        full_name = obj.employee.first_name + " " + obj.employee.last_name

        return full_name.strip() if full_name.strip() else obj.employee.email

    def get_leave_type_name(self, obj):
        return obj.leave_type.name if obj.leave_type else None

    def validate(self, data):
        start_date = data.get('start_date')
        end_date = data.get('end_date')
        leave_type = data.get('leave_type')
        document = data.get('document')

        if start_date and end_date:
            if end_date < start_date:
                raise serializers.ValidationError({
                    "end_date": "end_date cannot be before start_date."
                })

        if leave_type and not leave_type.is_active:
            raise serializers.ValidationError({
                "leave_type": f"'{leave_type.name}' is currently inactive."
            })

        if leave_type and leave_type.requires_document and not document:
            raise serializers.ValidationError({
                "document": f"A document is required for '{leave_type.name}'."
            })

        if start_date and end_date and leave_type:
            duration = (end_date - start_date).days + 1
            if duration > leave_type.max_days:
                raise serializers.ValidationError({
                    "end_date": (
                        f"Duration of {duration} days exceeds the maximum "
                        f"of {leave_type.max_days} days for '{leave_type.name}'."
                    )
                })

        return data

class LeaveStatusUpdateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Leave
        fields = ["status", "admin_remarks"]

    def validate_status(self, value):
        if value not in Leave.Status.values:
            raise serializers.ValidationError("Invalid status value.")
        return value 
